refactor(sql): Move playlist sql to sql directory
Move playlist sql to sql directory.
