+++ src/core/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(searchparser)
+add_subdirectory(sql)
 target_sources(MPax PRIVATE playcontent.h playcontent.cpp)
 target_sources(MPax PRIVATE playlist.h playlist.cpp)
 target_sources(MPax PRIVATE playlistjson.h playlistjson.cpp)
-target_sources(MPax PRIVATE playlistsql.h playlistsql.cpp)
 target_sources(MPax PRIVATE main.cpp)
