+++ src/core/sql/playlistsql.cpp
